Move solo tests to top tier
As these tests either never had, or no longer have other tests with them, there is no value in keeping them in their own folder.
